+++ docs/.vitepress/meta.ts
@@ -1,0 +1,14 @@
+import {version} from '../../package.json'
+
+// base info
+export const baseUrl = '/forguncy-guide/'
+export const siteTitle = '格言格语'
+export const siteLogo = ''
+export const siteKeywords = '低代码,活字格,forguncy,葡萄城,grapecity'
+export const siteDescription = '一站式企业级低代码开发平台最好的学习基地'
+
+// docs version
+export const docsVersion = version
+
+// social link
+export const github = 'https://github.com/PetrichorWandering/forguncy-guide'
