+++ docs/.vitepress/theme/index.ts
@@ -1,0 +1,9 @@
+import DefaultTheme from 'vitepress/theme'
+
+// custom CSS
+import './style/print.css'
+
+export default {
+  // Extending the Default Theme
+  ...DefaultTheme,
+}
