+++ docs/.vitepress/theme/style/print.css
@@ -1,0 +1,7 @@
+@media print {
+  .VPNav,
+  .VPLocalNav,
+  .VPDocFooter {
+    display: none !important;
+  }
+}
