+++ docs/lb-index.md
@@ -1,0 +1,20 @@
+---
+layout: home
+
+title: 格言格语
+titleTemplate: 活字格集群负载均衡方案
+
+hero:
+    name: 格言格语
+    text: 活字格集群负载均衡方案
+    actions:
+        - theme: brand
+          text: 开始学习 ->
+          link: /solution/load-balance/introduction
+        - theme: alt
+          text: 联系专家 🧑‍💻
+          link: "https://affim.baidu.com/unique_40629805/chat?siteId=18064822&userId=40629805&siteToken=017307e30728fbca74b59b122906b279"
+    image:
+        src: /lb-log.png
+        alt: 活字格集群
+---
