--- docs/solution/index.md
@@ -1,1 +1,0 @@
-# 解决方案一
