--- docs/solution/load-balance/high-availability.md
@@ -1,4 +1,0 @@
-# 高可用性集群
-
-本节介绍如何在生产环境中为活字格集群
-
